Properly get cygwin dir from sub-configure
When configuring subdirs of newlib, such as libc for example, the path
to the winsup directory is calculated relative to the current srcdir,
which only works at the top-level.
Fix the path to be relative to newlib's top-level srcdir.

newlib/ChangeLog:
configure.host: use $newlib_basedir to calculate cygwin_srcdir